[UPDATE] 로직 변경에 따른 메시지 기능 수정

+++ api/src/main/java/com/bouquet/api/message/dto/MessageRequest.java
@@ -9,7 +9,7 @@ public class MessageRequest {
     @AllArgsConstructor(access = AccessLevel.PRIVATE)
     public static class Create {
         private Long postId;
-        private Long iconId;
+        private int iconId;
         private String nickname;
         private String content;
         private int x;
